Remove "paste on set clipboard" feature
Now that content is pasted on Ctrl+v, the hack to send the PASTE keycode
(for Android >= 7 only) is obsolete. Remove it.

The shortcut itself can still be useful though, when Ctrl+v does not
paste (in Termux for example, it generates ^V, but pasting is still
possible via the context menu).

This reverts fc1dec027093ee8a2746a212db0c72a9d96d39a4.