Use VideoStreamer directly from ScreenEncoder
The Callbacks interface notifies new packets. In addition, the screen
encoder will need to write headers on start from its own thread (which
will be added later).

We could add a function onStart() instead, but for simplicity, just
remove the interface, which brings no value, and call the streamer
directly.

Refs 87972e2022686b1176bfaf0c678e703856c2b027