Fix an "expected expression" error
In C, a label can only be followed by a statement, not a declaration.
An error in `app/src/screen.c` violated this, and led to a build error
with an error message similar to the one below:

```
  ../app/src/screen.c:821:13: error: expected expression
              bool ok = sc_screen_init_size(screen);
              ^
  /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/clang/13.0.0/include/stdbool.h:15:14: note: expanded from macro 'bool'
  #define bool _Bool
               ^
  ../app/src/screen.c:822:18: error: use of undeclared identifier 'ok'
              if (!ok) {
                   ^
  2 errors generated.
```

This could be fixed by introducing a new block (or compound statement;
as is already being done in the next `case`). That is a statement.

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>