Avoid "negative" average buffering
Average buffering may not be negative.

Since dropping samples instantly reduces buffering, it is applied
immediately to the average value.

However, if an audio output issue causes the SDL callback to be called
too infrequently, samples will (silently) not be consumed as expected.
As this will cause samples to be dropped due to overbuffering, the
average value may drop below 0, which makes no sense.