Avoid "negative" average buffering
The assumption that underflow and overbuffering are caused by jitter
(and that the delay between the producer and consumer will becaught up)
does not always hold.

For example, if the consumer does not consume at the expected rate (the
SDL calback is not called often enough, which is an audio output issue),
many samples will be dropped due to overbuffering, decreasing the
average buffering indefinitely.

Prevent the average buffering to become negative to limit the
consequences of an unexpected behavior.