Add connected gamepads on start
Trigger SDL_CONTROLLERDEVICEADDED for all gamepads already connected
when scrcpy starts. We want to handle both the gamepads initially
connected and the gamepads connected while scrcpy is running.

It is a bit racy, because a device may be added after the controller
subsystem initialization, but before the call to list the joysticks. In
that case, a single device may in theory be detected as added twice.
This should be harmless though (the second initialization will just fail
without major impact).