Listen to display changed events
Replace RotationWatcher and DisplayFoldListener by a single
DisplayListener, which is notified whenever the display size or dpi
changes.

Still use the old version specifically for Android 14, where
DisplayListener may be broken (it is fixed in recent Android 14
upgrades), until we receive the first DisplayListener event (which
proves that it works).

Refs #4469 <https://github.com/Genymobile/scrcpy/pull/4469>

Co-authored-by: Simon Chan <1330321+yume-chan@users.noreply.github.com>