Add --otg-automatic-screenshot
With this option, scrcpy takes a screenshot automatically each time
mouse capture is exited in OTG mode.
This is particularly useful when OTG mode is used to enable USB
debugging on a phone with a broken screen.
The screenshot will display the mouse pointer that will indicate where
to move next.